Pipeline registers
Added majority of the pipeline reg banks, need to do flags, end of
writeback/intro to next stage, "next" stage needs to be heavily
redefined and removed from the control stage if applicable. Note a jump
is still feasible to perform in the decode stage due to lack of
accumulator. AFter this we need to do forwarding and hazard analysis.